requirements.txt
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.28
django-coverage==1.2.4
django-forms-bootstrap==3.1.0
django-storages==1.9.1
docutils==0.15.2
idna==2.9
jmespath==0.9.5
Pillow==7.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
stripe==2.45.0
urllib3==1.25.9
